Create fizzbuzzzuehlke.js
Given is a number which is not divisible by 3 or 5.
When the function called with this number.
Then the function should return the invoked number.

Given is a number which is divisible by 3. "fizz".

Given is a number which is divisible by 5. "buzz".

Given is a number which is divisible by 3 and 5. "fizzbuzz".